DepthLayer IsWaterCell() and IsLandCell() use single instead of integer
git-svn-id: https://sources.ecopath.org/svn/Ecopath/trunk/Ecopath6@11932 a725721c-0ed6-864e-a6a8-02c2a1988c9d

+++ Sources/EwECore/EcoSpace/Input Output/EcospaceBasemap/cEcospaceLayerDepth.vb
@@ -44,7 +44,7 @@ Public Class cEcospaceLayerDepth
     ''' -----------------------------------------------------------------------
     Public Function IsWaterCell(ByVal iRow As Integer, ByVal iCol As Integer) As Boolean
         If Not Me.ValidateCellPosition(iRow, iCol) Then Return False
-        Return CInt(Me.Cell(iRow, iCol)) > 0
+        Return CSng(Me.Cell(iRow, iCol)) > 0
     End Function
 
     ''' -----------------------------------------------------------------------
@@ -57,7 +57,7 @@ Public Class cEcospaceLayerDepth
     ''' -----------------------------------------------------------------------
     Public Function IsLandCell(ByVal iRow As Integer, ByVal iCol As Integer) As Boolean
         If Not Me.ValidateCellPosition(iRow, iCol) Then Return False
-        Return CInt(Me.Cell(iRow, iCol)) <= 0
+        Return CSng(Me.Cell(iRow, iCol)) <= 0
     End Function
 
     ''' -----------------------------------------------------------------------
